// csnfdockerdecorator/types.d.ts
/**
 * (C) Copyright IBM Corp. 2021.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export declare type Provider = {
    providerId: string;
    providerType: string;
    name: string;
}
export declare type Source = {
    sourceName: string;
    sourceId: string;
}
export declare type Event =  {
    name: string;
    guid: string;
    shortDescription: string;
    startTime: string;
    severity: string;
}
export declare type Resource = {
    identifier: string;
    type: string;
    name: string;
    region: string;
}
export declare type Decoration = {
    registry: string;
    namespace: string;
    image: string;
    tag: string;
    digests: string[];
    imageLastUpdated: string;
    description: string;
    starCount: number;
    pullCount: number;
}
export declare type CsnfEvent = {
    provider: Provider;
    source: Source;
    event: Event;
    resource: Resource;
    image: string;
    decoration: Decoration;
};